netconf-tutorial/pyproject.toml
[project]
name = "schema-driven-configuration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ncclient>=0.6.16",
    "pyyaml>=6.0.2",
]
